Implementend first draft of article edition page
