+++ web/admin-gui/src/router/index.ts
@@ -5,6 +5,7 @@ import Debug from '@/views/Debug.vue'
 import Login from '@/views/Login.vue'
 import Home from '@/views/Home.vue'
 import Articles from '@/views/Articles.vue'
+import Edition from '@/views/Edition.vue'
 
 const router = createRouter({
 	history: createWebHistory(import.meta.env.BASE_URL),
@@ -39,6 +40,14 @@ const router = createRouter({
 			component: Articles,
 			meta: {
 				title: 'GohCMS - Articles'
+			}
+		},
+		{
+			path: '/articles/edit/:articleID',
+			name: 'edition',
+			component: Edition,
+			meta: {
+				title: 'GohCMS - Edition'
 			}
 		},
 	]
