+++ web/admin-gui/src/views/Debug.vue
@@ -1,5 +1,4 @@
 <script setup lang="ts">
-import Editor from '@/components/Editor.vue';
 import { useAuthStore } from '@/stores/AuthStore';
 import { ref } from 'vue';
 
@@ -24,10 +23,7 @@ function ping() {
 </script>
 
 <template>
-	<div class="editor">
-		<Editor :show-default-data="true"></Editor>
-	</div>
-	<!-- <div class="inputs-group">
+	<div class="inputs-group">
 		<div>
 			<input type="text" placeholder="article ID" v-model="articleID">
 			<button class="button-primary" @click="getArticles(articleID)">get article</button>
@@ -35,7 +31,7 @@ function ping() {
 	</div>
 	<div class="buttons-group">
 		<button class="button-primary" @click="ping()">ping</button>
-	</div> -->
+	</div>
 </template>
 
 <style scoped>
